@@ -4,8 +4,7 @@
         "node": ">=14.0.0"
     },
     "scripts": {
-        "build": "webpack --mode production --config webpack.config.js",
-        "watch": "webpack --mode development --watch --config webpack.config.js"
+        "watch": "webpack --config webpack.config.js"
     },
     "author": "Scott Woodhams",
     "license": "MIT",
@@ -43,10 +42,11 @@
         "prettier": "^1.18.2",
         "source-map-loader": "3.0.0",
         "style-loader": "^1.1.3",
-        "typescript": "^4.3.4",
+        "ts-loader": "^9.2.6",
+        "typescript": "^4.5.4",
         "uxp-types": "^0.1.2",
         "webpack": "5.0.0",
-        "webpack-cli": "^3.3.2"
+        "webpack-cli": "^4.9.1"
     },
     "dependencies": {
         "react": "^17.0.2",
