@@ -5,9 +5,9 @@ import { ReadFromMetaData } from '../typescript/Metadata';
 import UILayerData from '../typescript/UILayerData';
 import TextDetails from "../components/TextDetails";
 import Spectrum, { Divider } from "react-uxp-spectrum";
-export type ActionPanelProps = { onExport: () => void };
+export type ActionPanelProps = { onExport: () => void, onSlice: () => void};
 
-export default function ActionPanel({ onExport }: ActionPanelProps) {
+export default function ActionPanel({ onExport, onSlice }: ActionPanelProps) {
 
   let emptyData = new UILayerData();
   const [metadata, setCurrentMeta] = useState(emptyData);
@@ -27,6 +27,10 @@ export default function ActionPanel({ onExport }: ActionPanelProps) {
     onExport();
   }
 
+  const Slice = () => {
+    onSlice();
+  }
+
   useEffect(() => {
     action.addNotificationListener(events, listener);
     return () => {
@@ -40,6 +44,7 @@ export default function ActionPanel({ onExport }: ActionPanelProps) {
       <Spectrum.ActionButton onClick={Export}>Export</Spectrum.ActionButton>
       <Divider size="large"/>
       <SliceRect rect={metadata.Bounds} slices={metadata.Slices} sliceType={metadata.SliceType} />
+      <Spectrum.ActionButton onClick={Slice}>Slice</Spectrum.ActionButton>
       <Divider size="large" />
       { metadata.TextDescriptor !== undefined && <TextDetails desc={metadata.TextDescriptor}/>}
     </div>
